# backend/openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "IExist API",
      "description": "IExist 서비스의 API 명세",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/api",
        "description": "API 기본 경로"
      }
    ],
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "schemas": {
        "Error": {
          "type": "object",
          "properties": {
            "status": {
              "type": "string",
              "example": "error"
            },
            "message": {
              "type": "string"
            }
          }
        },
...
        "User": {
          "type": "object",
          "properties": {
            "user_id": {
              "type": "integer"
            },
            "username": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "nickname": {
              "type": "string"
            },
            "profile_image_url": {
              "type": "string",
              "nullable": true
            },
            "theme_preference": {
              "type": "string",
              "enum": ["light", "dark", "system"]
            },
            "is_active": {
              "type": "boolean"
            }
          }
        },
...
        "Notification": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "user_id": {
              "type": "integer"
            },
            "content": {
              "type": "string"
            },
            "notification_type": {
              "type": "string",
              "enum": ["like", "comment", "challenge", "system"]
            },
            "related_id": {
              "type": "integer",
              "nullable": true
            },
            "is_read": {
              "type": "boolean"
            },
            "created_at": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
...
        "Pagination": {
          "type": "object",
          "properties": {
            "current_page": {
              "type": "integer"
            },
            "total_pages": {
              "type": "integer"
            },
            "total_count": {
              "type": "integer"
            },
            "has_next": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
...
      "/users/profile": {
        "get": {
          "tags": ["Users"],
          "summary": "사용자 프로필 조회",
          "responses": {
            "200": {
              "description": "프로필 조회 성공",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "data": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "인증 필요",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["Users"],
          "summary": "사용자 프로필 업데이트",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nickname": {
                      "type": "string",
                      "minLength": 2,
                      "maxLength": 50
                    },
                    "theme_preference": {
                      "type": "string",
                      "enum": ["light", "dark", "system"]
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "프로필 업데이트 성공",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "message": {
                        "type": "string",
                        "example": "프로필이 성공적으로 업데이트되었습니다."
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "nickname": {
                            "type": "string"
                          },
                          "theme_preference": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "잘못된 요청",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "401": {
              "description": "인증 필요",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        }
      },
...
      "/notifications": {
        "get": {
          "tags": ["Notifications"],
          "summary": "알림 목록 조회",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "schema": {
                "type": "integer",
                "default": 1
              }
            },
            {
              "name": "limit",
              "in": "query",
              "schema": {
                "type": "integer",
                "default": 10
              }
            },
            {
              "name": "type",
              "in": "query",
              "schema": {
                "type": "string",
                "enum": ["like", "comment", "challenge", "system"]
              }
            },
            {
              "name": "is_read",
              "in": "query",
              "schema": {
                "type": "string",
                "enum": ["true", "false"]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "알림 목록 조회 성공",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "notifications": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/Notification"
                            }
                          },
                          "pagination": {
                            "$ref": "#/components/schemas/Pagination"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
                "description": "인증 필요",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Error"
                    }
                  }
                }
              }
            }
          }
        },
...
      }
    }